After-merge update + rename the error type + move to another file
